Start writing custom unmarshallers for config.
This will be used to do much fancier parsing and more optional values.  At the same time, inputs and outputs are going to be restructured to be maps -- while it does make sense to just *list* them so that they're conceptually appendable, in practice this has turned out to be a drag: when constructing pipelines of formulas, we continuously need a way to refer to inputs and outputs by name so that a third party can work with the API reasonably.  So, we're gonna make that happen.

Signed-off-by: Eric Myhre <hash@exultant.us>