Ahum.  We're reaching the point were we have to start stripping the old json decoder hints.
This is slightly frustrating because I'd like to keep them there as a form of documentation... but it's time that we write, well, *Documentation* rather than relying on godoc of tags on structs.

AMEND: UGH this PR is now officially off the deep end for passing-at-every-commit.  We're going to keep recording historical snapshots, but... even though this commit fixes all CI checks, it's semantically breaking: the repeat-thyself script is now not parsing the filters, and will quietly fail to be repeatable.  (We should add a metaCI step for that!)

Signed-off-by: Eric Myhre <hash@exultant.us>