Remove attempt to autoderive goroot.
I added this because some people had bad first-time experiences unzipping a go toolchain and expecting it to work; now I'm getting other reports that other install mechanisms splay scripts and symlinks in such a way that this in turn breaks them out-of-box... I give up.  I'm deeply disappointed by tools that don't automatically find their own butts when run from a non-compile-time-hardcoded directory, and asking people to patch their bashrc env defaults is a user-hostile approach that smacks of the most ancient cultural mistakes of unix dark ages which I'd thought we could all overcome together in this decade, but you know what, this is not my battle.

Signed-off-by: Eric Myhre <hash@exultant.us>