Use local repo in repeat-thyself if available :sparkling_heart:
Note that though we've finally got config support in order, for this to work correctly it does still require correct support per transmat.  In the git case, we'll correctly failover down the list if a repo is unreachable, but there's not yet support for failover if a repo exists but is missing the requested data (this is a little tricky to ask in an efficient manner with git).

Signed-off-by: Eric Myhre <hash@exultant.us>