Use filter defaults in repeat-thyself :sparkling_heart:
The defaults now "do the right thing" and are appropriate to a typical reproducibility environment (and have for a while actually; this change is slow in coming).

It was nice to have an example of using filters handy, but at the same time, at this point I'd just as druther not have the repeatr build be the demonstration case for eccentricities.  (Also, in the most literal sense, opening those tars and getting the mtime=0 on my own disk has been getting annoying because several tools I use mistake that for a nil value, so I'm just following a path of headache minimization.))

And yes, all commits touching repeat-thyself must end in a sparkling heart emoji.

Signed-off-by: Eric Myhre <hash@exultant.us>