Move/reinforce boundary between placefile and tar.
Explicitly list all the supported modes in the tar code; we'd also be doing a mapping here if we had used a separate set of consts for our internal representation (which, really, we should have).  Change unexpected modes in placefile into a programmer error: they should be filtered before getting that deep.

Signed-off-by: Eric Myhre <hash@exultant.us>