Add concept of BareMount to assembly.
Start passing the those flags for mounts directly to host filesystems down towards the assembler.

Add more documentation to Assembler and AssemblerPart to describe this boundary: this layer is for expressing the intent of what you want where and how closely connected it should be; below is for worrying about how to get it there.  This has always been the rough intention, but I sort of had to re-convince myself "yeah, there's a Reason we don't want COW placers to be seen all the way up in executor/util/provision.go" so it's evidentally time to put that in writing.

Signed-off-by: Eric Myhre <hash@exultant.us>