Break ground on integrating opencontainers runc.
"Runc"'s upstream sources are at https://github.com/opencontainers/runc ; we'll be interacting with it as a plugin via exec'ing it.

This executor will be... well, speaking plainly for a moment, much the same as the nsinit one we've already had for a while now: though this is ostensibly a different backing organization, most of the code is flat out the same if you look past the name and related various publicity stunts.  I'm not questioning any of the reasoning behind this; runc appears to be The One going forward, and the previous nsinit product in libcontainer is deprecated and unsupported, so, here we go: new API to deal with, new package in repeatr.

This commit contains a snapshot of the default config emitted by `runc spec` built from git source hash=902c012e85cdae6bb68d8c7a0df69a42f818ce96 (the tip of master as of last night).  We'll be building permutations out from here.

Signed-off-by: Eric Myhre <hash@exultant.us>