Template config, wrap runc invocation, tests.
Tests will FAIL as yet: this is going to be one of those PRs that's big enough that I'm going to start tracking diffs before anything well, works.

Runc is currently expected to be on your path by magic.  Before landing this branch, we're going to build a plugin acquisition framework based on reusing repeatr's already existing IO components, and use that to get assets like the runc tool in place without an out-of-band fussy install process.

Signed-off-by: Eric Myhre <hash@exultant.us>