Pass full rootfs path into runc config.
It does not evaluate the path relative to the config or runtime files, so we have to push the whole thing into the template instead.

Note that this is still entirely tolerant of relative paths (perhaps somewhat remarkably).  CWD of the master repeatr process is simply never changed.  (There was a while where uncommitted work pursued setting a different cwd for the runc commands, but that didn't seem to solve any problems, so the approach was dropped.)

Signed-off-by: Eric Myhre <hash@exultant.us>