slight detour: make an asset fetcher and manager.
This is primitive looking, but really, the vast majority of the work is done already by the transmats.  It's kinda like, ya know, we kind happen to have this whole data transport with integrity guarantees and CAS caching thing going on.  Just by, like, coincidence, maaan.

Unfortunately, it's kinda tricky/useless to write tests for this that aren't full networked integration tests, so coverage shall be... uh, low.

Signed-off-by: Eric Myhre <hash@exultant.us>