Legitimate hash for a runc asset.
Note well that though this was produced from a formula, runc builds invoke gcc in a way that is not (out-of-box, anyway, and I haven't researched how to patch it yet) repeatable.  So, you're trusting me and my one-time compiler invocation on this one.  Hopefully we can move the needle further on that in the future: ideally we'd have a repeatable formula for production of any/ever plugin checked into the repeatr source repo.

Signed-off-by: Eric Myhre <hash@exultant.us>