Use runc asset plugin.
Tests in CI can now stop failing from not having a system runc.  Because why would we rely on a "system" dependency? :D

They'll still fail for a variety of other reasons, of course.  Namely, error handling paths are still not up to par with the spec.  But a decent amount should go green now.

Bump the version number in the config file.  This new build demands it.  No further comment.

Unfortunately this does *NOT* yet well solve the problem for development-on-an-airplane, a subject near and dear to my heart.  Since the assets cache is per basedir, and we quite intentionally make a new one of those per test instance, there's a bit of a bootstrapping snafu there.  That'll require future work on the plugins asset fetching system.

Signed-off-by: Eric Myhre <hash@exultant.us>