Sidestep for a moment: reparent exec errs, cwd err
I was about to add CWD error handling to runc, and realizing that didn't have a type of its own bothered me too much.

Also it's clearly wrong to treat no-such-command as a "should not occur in normal operation" error, so, fixed that.

Updated tests asserting on that error type, fixed chroot executor to match, and added another test for the "it's there but not a file" case.  (Nsinit is unchanged because it already never supported this correctly; and at this point, we're full steam on runc which when done should fully deprecate nsinit.)

Signed-off-by: Eric Myhre <hash@exultant.us>