Handle command-not-found in runc.  With todos.
Further auditing of the runc source has turned up a *bunch* of locations where critical errors occur and are not reported in a way that can be separated from valid user content.  More patches will be required upstream to address this safely and correctly.

Also, even relative outlier error cases aside, apparently things like command-not-found get written to stderr -- yes, also not separated from valid user content.  In this case we can at least *tell* and do something about it because it's *also* emitted on the control band, and fortunately these errors coincide with an absolute lack of possibly valid user output, so we can safely toss the redundant noise out the window, but the whole stirfry is just frankly obnoxious.  Control channels, people.  They're like, a concept.

Signed-off-by: Eric Myhre <hash@exultant.us>