Update default base image for interactive mode.
(This one has 'make' and some other nicities installed.  As before, (*extremely*) regretably, there's no audit chain of reproducible formulae for this asset, because none of the apt-based tools we're using are well oriented for that.  If someone wants to PR a way to recursively "pin" apt deps... please do; I've never seen it done.)

Signed-off-by: Eric Myhre <hash@exultant.us>