Fix one (1) ton of dir transmat legacy behaviors:
- data locations are now URIs!  (like everyone else, finally)
- data locations can be content-addressible!  (like everyone else, finally)
- multiple commit locations for an output are supported.  (like everyone else.)
- arbitrarily deep directories will not be created for output.  (like everyone else.  it's sanity preservation in the event of fat-fingering.)
- commits are transactional!  (tmpdirs are used correctly, and atomically moved into place.)
  - except if you're using non-CA mode.  because that's impossible within the limits of posix filesystem's atomicity primitives.  As if you needed yet another reason to go full-on CA...

Overall, this goes a long way towards bringing the dir transmat back in line with "what you'd expect" if you're accustomed to the config of any of the other supported core transmats, and makes it easier to swap out (e.g. s/tar/dir/ for the type, and it'll work now... as you'd expect).

Signed-off-by: Eric Myhre <hash@exultant.us>