Add `repeatr unpack` command.
This can be used to just deliver a filesystem somewhere on a host machine (without any processing).  This is useful if you want to hand files off to some other tooling at the end of a pipeline of repeatr processes, or if you just want to eyeball 'em to see what happened.

Signed-off-by: Eric Myhre <hash@exultant.us>