Make tar transmat report DataDNE all the way up.
This is the kind of behavior that we'll see be handled much more clearly as we move towards the "warehouse" as an abstraction in code as the dir transmat is already leading off; this is just a patch to the current implementation without much restructuring effort.

Report DNE briefly in relevant cli commands.

Signed-off-by: Eric Myhre <hash@exultant.us>