Update version of ugorji codecs.
This is gives us working codec.Selfer support and MapBySlice, which together mean we can buy per-struct decoding hooks that are mostly automatic, but also make things deterministically ordered on maps, which is a major requirement for sanity.

Note this also isn't quite the latest version, and that's on purpose: there are to some regressions on MapBySlice behavior I just discovered and reported as issue 124 on that project; this is the last working version on that feature until those get fixed.

Signed-off-by: Eric Myhre <hash@exultant.us>