Create def.InputGroup: orient around maps.
Previously, we were using `[]Input` as the format of note: that "worked", but we spent more time flipping it to a map than not, so: we're going to have maps be the canonical format, and use ugorji codec's MapBySlice technique to serialize them in deterministic order.

This new InputGroup type will need to replace the input slices in `def.Formula` before we're done, but I'm not tackling that quite yet because that change will hit a looot of other code, some of it quite nontrivally.

Signed-off-by: Eric Myhre <hash@exultant.us>