Implement codec for OutputGroup.
This is all but entirely copypasta of InputGroup (and I refactored names of some of the map->slice code in inputgroup to be literally copypasta) -- on some levels this is okay; on others, frustating.  The two *are* deeply similar, but making them share more code by field embedding has been considered before and rejected as currently not sufficient ROI; more frustratingly, the map->slice code... currently, I'm aware of no way to do it within the confines of golang without copy paste (this is a scenario where one might choose to invoke go-generate to template things, if it got much further out of hand).

Signed-off-by: Eric Myhre <hash@exultant.us>