Re-apply the mountpath defaults in IG and OG codec
Tweak test to prance around the mutation behavior there because that's not what that test is intended to cover (though we should back up and add another test that does).

We'll remove the `Name` field from inputs and outputs entirely when we throw the switch on cutting `Formula` over to `InputGroup` and `OutputGroup`.  Which... is possibly soon now.  I'm almost out of incremental changes.

Signed-off-by: Eric Myhre <hash@exultant.us>