Here's the big one: cut over, and use In|OutGroups
Most of the impacted code is confined to the executor provisioning util code and some test fixtures, which was kind of awesome.  (I was afraid this was gonna be way more painful.)

Add a handy new onelinear to parse from yaml strings.  Reorient current stringy tests to use that, as well as the config parser in the main commands.

Discard all the old hand-rolled unmarshalling.  Which is mostly a relief.  Somewhat alarmingly, this also means we throw away all the tests we had based on struct comparisons.  Since we don't have that intermediate form anymore... well, that strategy of testing can't work for us anymore.  This is okay, because we could more or less never deal with the ordering issues on that approach.  We'll just need that many more full string tests.

There's some lingering oddities around the fact several structure switched to pointers instead of plain structs.  Those will continue to drift towards pointers as they come up.  The landmine of slices and maps inside plain structs not being subject to meaningful copy-on-call makes that feature useless as a defensive coding mechanism, and the fact that the memcopies can actually register on performance is just beleaguering: I'm giving up the ghost and quietly switching to pointers by default in basically all situations from now on aside extreme compelling special circumstances.

Signed-off-by: Eric Myhre <hash@exultant.us>