Teach git transmat to emit version of git tool.
We're still relying on the host to have an install of git -- unfortunately.

And this is (surprise, surprise!) currently biting us in CI!  Amazing the irony, isn't it.  Travis switched over some of their server infra lately, and in the process... quietly changed the version of git they ship in their machine images.  Which, lo, breaks repeatr integration, because -- and get this, of all the things -- because this version of git, it doesn't like dirs existing before cloning into them, so it thunks hard because of course we're creating a tmpdir before pouring into it... Yeesh.

So.  This is a case study in why referencing incredibly, resolutely specific versions of your dependencies is a good idea!  Doesn't matter what anyone things a minor version update is.  The goofiest shit can wreck your build.  Pin.  It.  All.

Or your CI randomly goes red on a saturday night for reasons entirely out of control and unrelated to your work when you really just wanna land this 2ksloc PR and get on with your life.

Signed-off-by: Eric Myhre <hash@exultant.us>