Travis: update git.
A better fix for this is to build and package a deterministic, pinned version of git and bundle it to repeatr, like we do with runc.  But today, well, today I want CI running again so I can CI all the commits along the way to getting that long-term solution done right.

Signed-off-by: Eric Myhre <hash@exultant.us>