Use runc as the executor in the twerk subcommand.
Runc gives us out-of-box support for tty emulation, which is a vastly nicer experience.  (That feature could be implemented for the chroot executor as well, but it's easier to just switch to runc and toggle it on.  We may want to back and implement it in chroot as a fallback too at some point in the future.)

Signed-off-by: Eric Myhre <hash@exultant.us>