Begin extracting 'warehouse' abstraction in git transmat.
Following the same outline as in the dir transmat.

The trigger for this is realizing that travis tests on merge branches are failing to fetch travis's synthetic merge commit for a very interesting reason: because travis has a *shallow clone*, and you can't pull from those.  I want that remarkable eccentricity to come out in logs, because otherwise it's quite a headscratcher.

Signed-off-by: Eric Myhre <hash@exultant.us>