Enhance error reporting in git repo ping.
The errors returned when failing to contact a remote repo now contain a little more helpful of strings.  Additonal errors I haven't seen before might not be well captured here, but capturing and summaring all possible outputs of a command is always going to be climbing Mount Olympus as Sisyphus; this is better than before.

Signed-off-by: Eric Myhre <hash@exultant.us>