Okay, finally knock out the travis problem.
There's nothing we can do to make repeat-thyself runnable in the travis PR environment: they created a synthetic commit, marked that the head, and yet we can't clone it from the local repo even knowing that because it's a shallow, and git (at least any version we're currently interacting with) fiat refuses to clone from shallow repos, regardless of whether or not the requisite objects are available (whee).  We *could* attempt to shallow-clone ourselves as well, but that proceeds down a messy path -- git's remote API also currently (natch, this changes in *extremely* recent versions, and only with explicit cooperation of the server) doesn't let us get any clue of how back we have to clone in order to see our object of interest.

hashtag gitproblems.  Everything makes consistent sense; it just all resolves to an ontological universe that's in the shape of "bummer".

Signed-off-by: Eric Myhre <hash@exultant.us>