Move first draft derive code to new actor model.
Flipped catalogs to a concrete type as described in plans a few commit messages ago.

This first actor implementation is using a strategy of selecting over a couple different work triggering events... and a channel that reports old stuff without triggers is simply another one of those: this model solves the problem of how to react to fresh inputs in realtime while still backfilling (whenever time permits) for the possibility of missed events when the actor wasn't yet alive and registered for observables.  This pattern should work well for expressing that basic priority and scale freely to more kinds of triggers.

Leaving lots of todos strew about on the insides of the knowledge base components themselves, still; just trying to get all the pieces to line up so they can be coordinated efficiently and in race-free ways.

Signed-off-by: Eric Myhre <hash@exultant.us>