Fleshing out knowledge base implementation.
Starting to try to draft tests on this, which is good, because not everything in the current foreman draft is well separated for testability, and that is a major code smell now correctly highlighted.

There will be more knowledge base implementations in the future (I'm thinking database-backed storage will be desirable at some scale), but haven't switched to an interface yet.

Signed-off-by: Eric Myhre <hash@exultant.us>