Refactor foreman to have stages -- prep for tests.
These might not be the final boundaries of functions (it's still pretty high granularity), and also the event pumping and the formula evocation may end up in separate goroutines later.  The key point for now is that some of the state is getting exposed, and the functions that mutate it can be called in single steps so we can observe them and test via prying open the state.  Definitely not black box testing, so the test specs won't be reusable, but that's okay; different actors have the option of such radically different opinions about work planning that I'm not sure tests would ever be reusable anyway.

Signed-off-by: Eric Myhre <hash@exultant.us>