Do the simplest thing to get some benches.
Reusing goconvey stuff for benchmarking is a little... hectic.  The output is currently an unacceptable mess.  And I am not happy about that `atoi` hack, but it's necessary to keep order when that loop happens -- same reasonsing as why we wrote `testutil.AdditionalDescription`.

Not sure what the long term answers to that is going to be; for now, rolling with it, and conserving any aggressive choices until after we do some pprof or such to see if goconvey also imposes an unacceptable amount of overhead for e.g. its logging and context finding and whatnot.  If we have to start mega-parameterizing tests with logging io.Writer refs and so on, I'm going to be... well, sort of sad, for obvious reasons.

Signed-off-by: Eric Myhre <hash@exultant.us>