Basic tests for nil-executor deterministic mode.
Fix that executor to not mutate the formula it's handed in.  I think that's going to be part of the standard behavior going forward; implementations will be adjusted to that as they're found.  Meanwhile, the sanity check in the "stage2" ID function is pretty useful at catching BS.

Also, wildly enough, this is finally the first time we're seeing use of the `conjecture` flag *matter*.  So that's fun.  (Not exhaustively testing it here, though: it's coincidentally relevant to the way nil-det mode *happens* to be implemented; it should really be covered by its own suite of tests in the model package.  Will probably do that post-refactor of bringing all the existing `def` code into `model` (aka, at least one more PR down the road from the present).)

Signed-off-by: Eric Myhre <hash@exultant.us>