Regroup foreman tests.
All of these are for testing the foreman's response to existing and incoming catalog updates given some other state of commissions.  So say that.

I was originally thinking of moving on to add more tests with closing the loop here, but... that may not actually be the right API surface.  Release publishing actors should *not* get entangled with this work planner implementation.  Which means testing full-circle work stuff should come later, in another package for integration testing that pulls both hemispheres of the system together.

Signed-off-by: Eric Myhre <hash@exultant.us>