Mount the cwd when twerking.
This is usually "do what I mean" when I'm doing horse-around development.  Writable by default is also a little aggressive, but remember, twerk is *meant* to make you feel embarassed by using it, and is not the paragon of safe defaults.

Permissions are a little interesting.  UID 0 in a runc container as currently configured does *not* have permission to do a dang thing to your filesystem, unless it's already 777 or owned by uid 0.  So that limits utility somewhat sharply.  This could be resolved by switching to the invoking UID (or SUDO_USER's uid?), but I currently don't know where I would hang that on the config (formula.action.escapes?  it *really* shouldn't be used in "normal" operation.  maybe a new set of execution config out of band of formulas entirely is in order?)

Signed-off-by: Eric Myhre <hash@exultant.us>