Revise keep policies on work-in-progress formulas.
(Mostly, dropping comments from previous design concepts, because on further reflection there's a reason those todos seemed awkward to implement: they implied some nasty semiotic flaws down the road.)

Managing this whole merry-go-round at scale requires semantics we can use for garbage-collection.  And I'd rather not have that end up requiring fudge factors based on timestamps to avoid discarding WIP stuff.  So: GC is going to be based on strong references (because of course it is), we're going to start planning for that now, and the current obvious roots of interest are anything that made it into a relase catalog.  As a result, if you need state caching for WIP stuff, do it yourself; and inter-actor WIP message passing is currently seen as unspecified (because I don't want to deal with putting a "lease" pattern in the KB; trying to define sharing for those would be a snarl).

Signed-off-by: Eric Myhre <hash@exultant.us>