Foreman plan queues now much more structural; leases; keep commission info with plans.
Commission info is retained for the whole trip because we want to refer back to it later in the cycle when plotting out whether and where to publish releases.

Lease semantics should make it pretty easy to parallelize executors from here.  (Though we may not end up getting much mileage out of cancels until we get to either farming or a crash-resistant plan queue.)

Signed-off-by: Eric Myhre <hash@exultant.us>