First full step of evokation! :fire:
This is the first time we've tested a new catalog appearing in the knowledge base entirely as a result of formulas planned from commission+catalog.  Woohoo!

(Well, we're still pushing the foreman through its paces manually for test purposes (and yes, that's concealing a bug where as currently written the planner part will block the evoker), but still.)

Also, add sort interface for catalog IDs.  I seem to be needing sortables with increasing frequency just to deal with testing.  For the $n$'th time, I'd trade a sizable brick of gold for order-preserving maps.

Signed-off-by: Eric Myhre <hash@exultant.us>