Start work on minimum-viable environments.
In packaging more things and attempting to run with minimal perms, there's... it's possible for there to be a lot of friction around basics.  This is starting work to address that, while also making it easier to configure permissions dropping out of box (and hopefully, make that a reasonable default).

Signed-off-by: Eric Myhre <hash@exultant.us>