Defaults and MkCradle are separate phases.
The latter is to be disablable since it causes fs mutations, and we sometimes might want to explicitly avoid such inheritable things, etc.  Defaulting env vars might sometimes be unwanted, but I don't really anticipate a lot of impetus to do so, and it's already easily overriden within existing config.

Signed-off-by: Eric Myhre <hash@exultant.us>