ensureWorkingDir can mostly shell out to fs.MkdirAllWithAttribs, happily.
At this point, we need to map from the Policy info to a UID to define some of those fields.  I'm hesitant to go too far on that right away (YAGNI senses are tingling hard), but otoh, I'm pretty sure we're going to end up having a parser and writer for passwd files at some point shortly (even if that gets moved behind the cuttoff for this first PR, sadly I'm pretty sure it's coming -- I recall, for one random but certainly not rare or unique example, real nasty behavior in python when it finds your uid doesn't easily map to a name, which I don't think is a pain we should allow to be associated with low-priv operation)... and that means we might wanna just have a struct right now for representing all the smutchz necessary for that.

Going to take a middle road and have a struct right off the bat, at least.  Perhaps populate more fields later on-demand.

Signed-off-by: Eric Myhre <hash@exultant.us>