Add capability dropping to runc per Policy.
Not... tested.  I'm not entirely sure how to tackle this, tbh.  I think I'll have to slowly accumulate a corpus of "interesting" commands to run inside the full container to probe the boundaries of the security.  This will be a big and ongoing project.  (Possibly an entirely separate binary -- capsprobe?  ...Oh, I guess I could use capsh for now.  But I kinda want to get something that actually *tries* to do the_things and then asserts that they error exactly as much as expected, because that's more topically on point to the actual security of it all.)

Signed-off-by: Eric Myhre <hash@exultant.us>