Drop a cap from Sysad Policy because unsupported.
Runc is erroring out and calling CAP_AUDIT_READ unknown.  I'm somewhat confused by this, because it *is* documented.  Maybe it's doing detection of my system's available caps (I don't have a sufficiently recent kernel on this host to have this particular cap), but I haven't seen such code yet...

Commenting it out and moving on for the moment because I don't anticipate dropping CAP_AUDIT_READ even on systems that support it to be problematic.

Signed-off-by: Eric Myhre <hash@exultant.us>