Priv dropping for chroot executor!  + spec check.
This involves, alarmingly, upping the Policy for the repeat-thyself formula to have slightly more than the routine.  Why?  It's covered by our acceptance tests... and it's using the git transmat.  The git transmat is currently operating with more permissions than it should, and also leaving ownership of the checkout set to uid=0:gid=0, which lands this script in hot water unless it can write there.  We'll need to address this asap.

Signed-off-by: Eric Myhre <hash@exultant.us>