Fix git transmat perms.
The default perms for a transmat system with no inherent concept of perms and no filters configured should be 1000/1000.  Also, yeah, we should just plain be running git with far few privs.

This always should have been the case -- enabling priv-dropping by default is now shining a light directly on it (acceptance tests started failing!).  Yay safe defaults: calling BS where BS should indeed be called.

Aside: I continue to wish that it's reasonable to run the transmats themselves in sandboxes (anything outside of simple tar usage, at least; that one we're fairly willing to be confident of the code quality on)... and yet I continue to not do that, because it seems valuable to design transmats to work alone, moving things on and off the host filesystem in save/load subcommands, which should not themselves necessarily require priviledges or even container support, and that means continuing to refrain from using containers for the transmats.  Maybe someday there'll be a happier middle road in the future that gives us more isolation to the transmats without onerous requirements.  Or we'll design a way to make containers optional such that they provide failsafes when present while avoiding behavioral changes so that the transmats still work the same uncontained (just without the failsafes)... which come to think of it, actually sounds possible.  Future work.

Signed-off-by: Eric Myhre <hash@exultant.us>