Make git content owned by min-priv Routine Policy.
Unfortunately... This is backing out on getting git content permissions right by actually running git as a less privileged user in the first place.  That *would* be desirable, but see the comments in the commit body for an explanation of the current state of defeat in detail.  This subject will need to be revisited with significantly more firepower in the future.

On the plus side, with this change, the ownership and bits exposed by the git transmat to the inside of containers is not just correct, but correct in such a way that the less excellent internal parts are hidden... so when we *do* revisit this to make better engineered and lower-priviledged git transmat guts, it should be a completely transparent, not-at-all-breaking change.

Signed-off-by: Eric Myhre <hash@exultant.us>