Add changelog.
Adults keep changelogs.  I am an adult.  Startiiiing.... NAO

I was kinda wondering if it's time to throw repeatr a birthday party.  But it seems that moment has already slipped past :(

Version numbers for older stuff is fairly arbitrary.  If I followed the letter of the sem-ver law and only incremented the pre-1.0 "major" digit (aka the "minor") when a change was breaking, that number would remain *very* low indeed, so I'm simply upping it freely.

Signed-off-by: Eric Myhre <hash@exultant.us>