Support env var insertion directly.
Tempfiles and shenanigans shouldn't be needed if someone wants to do something this simple.  So, here it is.

Co-opted the `-e` flag.  It makes more sense for that to be for env vars.  Other tools use `-e` for that, and -- hopefully -- setting a non-default executor is going to be rather unusual.

Signed-off-by: Eric Myhre <hash@exultant.us>