Report exit code in outputs when using `repeatr run`.
The exit code has simply been added to the existing output group, under a well-known name.

It's possible this may end up in more formulas in general or in a special field all its own rather than piggybacking with the other outputs; we'll see how that goes in the future, and probably revist it on the other side of the next review&refactor pass to `BasicJob` stuff.

Signed-off-by: Eric Myhre <hash@exultant.us>