Remove vacuous test.
It's apparent this isn't doing anything, both from cursory review, and because if it was doing anything it should be broken right now.  Furthermore this is mostly/better covered by the demo and other acceptance tests.

Unit tests here aren't a *bad* idea, but this is such a false start it will be better to start over.

Signed-off-by: Eric Myhre <hash@exultant.us>