Document public release download locations.  And provide example bootstrap script.
(n.b. I *know* there's a lot of these, like IPFS's amusingly-named "hashpipe" for example; yes, I *know* there's an xkcd about the 15th standard; this one is a bash script you can embed in a project without a fuss.  There are many {scripts|copypasta} like it but this one is mine.)

Signed-off-by: Eric Myhre <hash@exultant.us>