Move somewhen to lib/fs.
Fixes the fact that a lib package was importing from other parts of the repeatr apis, which is supposed to be a no-no.

There are still more references to this floating around than I realized.  They should probably be broken down into different vars for "why", for starters.  Why bother?  Well, some places don't actually do well with a real epoch timestamp.  (I have file sync tools that get maaaaad and refuse to believe you if you have files with epoch properties, for example.  Which is a "stupid" problem, to be sure, but a real thing that's happening in the real world, sadly.)

Signed-off-by: Eric Myhre <hash@exultant.us>