Propose a RunRecord struct; this requires a Ware type; and add the first suggestion of a verb to the act package.
Previous design notes indicated having a "stage3" formula that was the repetition of the stage2 (aka, what's actually been called "formula" all this time), with the blanks for output hash filled in.  I'm kind of drifting away from that idea.  It might make sense to pastebin such a stage three formula as a build-instructions-and-attestation all-in-one, but for any other schema and situation it's just remarkably redundant.  So, instead, I'm thinking of pivoting towards these RunRecord types as the default way of keeping notes.

The "Ware" type is partially copied (well, mostly the comment, even) from the core/model package.  This pair is something that probably should've emerged a long time ago, and should also become an embed in the def.Input and def.Output types (in fact, it's been on the whiteboard since the early days, but at the time I tossed it out under "YAGNI" abstraction; it's become more clear over time that the SKU-like concept is actually the most recurrent one, especially as we move towards more abstractions around the catalog system and warehouse management).

Copies from the model package are indeed a code smell... my current plan is to scoop things out of core/model as api/act gets fleshed out, and then at a later date core/model can be revisted and swept out.  Hopefully at that point things look a bit different and most of the foreman actor can be implemented against act/api intead of linking executors directly, etc; this should make for prettier code, and also be much more ready for farming out to a cluster (the 'act' boundaries will be much more readily proxiable than the executor interface).

Signed-off-by: Eric Myhre <hash@exultant.us>