Take a crack at a 'mirror' func interface, and introduce a 'library' struct to help describe the desired states.
Toss a Catalog ID type back into the fray, but still less than 100% sure what we'll want from that.  As with the previous iteration on these concepts (still under the model package), there was an allusion to SFS-style private-key-backed claimery.  I'm not sure I want to make that a *requirement*, because privkey management in tests and demos and local operation tends to produce absurd amounts of frictions if not carefully managed, but maybe something to avoid purely human names without a gathering authory will be a good idea...?  I dunno.  TDB.

Signed-off-by: Eric Myhre <hash@exultant.us>