Drop schedulers.
I think the amount of green in this diff to replace them testifies to how clearly useless that abstraction was.

Also, remove one log line about "job started" that had *nothing* to do with *anything* aside from heart palpitations of the scheduler.  That message has mislead for the last time.

Signed-off-by: Eric Myhre <hash@exultant.us>